move tools into tests and remove conditional import to allow types to infer
