+++ tests/core/test_factories.py
@@ -1,9 +1,10 @@
-from eth_keys import (
-    keys,
-)
-from eth_keys.tools.factories import (
+from tools.factories import (
     PrivateKeyFactory,
     PublicKeyFactory,
+)
+
+from eth_keys import (
+    keys,
 )
